+++ library/android-mvc-test/src/androidTest/java/com/shipdream/lib/android/mvc/BaseTestCase.java
@@ -239,14 +239,15 @@ public abstract class BaseTestCase<T extends TestActivity> extends ActivityInstr
                 @Override
                 public void run() {
                     Mvc.graph().release(BaseTestCase.this);
-                    try {
-                        Mvc.graph().getRootComponent().getCache().clear();
-                        Mvc.graph().getRootComponent().detach(component);
-                    } catch (Component.MismatchDetachException e) {
-                        e.printStackTrace();
-                    }
                 }
             });
+        }
+
+        try {
+            Mvc.graph().getRootComponent().getCache().clear();
+            Mvc.graph().getRootComponent().detach(component);
+        } catch (Component.MismatchDetachException e) {
+            e.printStackTrace();
         }
 
         super.tearDown();