+++ library/android-mvc-test/src/main/java/com/shipdream/lib/android/mvc/view/injection/controller/ControllerB.java
@@ -16,13 +16,13 @@
 
 package com.shipdream.lib.android.mvc.view.injection.controller;
 
-import com.shipdream.lib.android.mvc.Controller;
+import com.shipdream.lib.android.mvc.FragmentController;
 import com.shipdream.lib.android.mvc.UiView;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class ControllerB extends Controller<ControllerB.Model, UiView> {
+public class ControllerB extends FragmentController<ControllerB.Model, UiView> {
     public static class Model {
         private List<String> tags;